added the login and registration functionlity

+++ src/component/Register.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
+import styles from "./RegisterModule.module.scss";
 
 const Register = () => {
   const navigate = useNavigate();
@@ -14,10 +15,20 @@ const Register = () => {
       password: e.target[2].value,
       confirmPassword: e.target[3].value,
     };
+
     console.log(user);
     let users = localStorage.getItem("users")
       ? JSON.parse(localStorage.getItem("users"))
       : [];
+    // find if the user already exist
+    const alreadyRegisterUser = users.some(
+      (item) => item.email.toLowerCase() === user.email.toLowerCase()
+    );
+
+    if (alreadyRegisterUser) {
+      alert("User already exist.");
+      return;
+    }
     users.push(user);
     localStorage.setItem("users", JSON.stringify(users));
 
@@ -26,16 +37,33 @@ const Register = () => {
     navigate("/register-success");
   };
   return (
-    <div>
-      <h2>Register</h2>
-      <form onSubmit={registerOnSubmit}>
-        Full Name : <input type="text" name="fullname" id="fullname" />
-        Email : <input type="email" name="email" id="email" />
-        Password : <input type="password" name="password" id="password" />
-        Confirm Password :{" "}
-        <input type="password" name="confirmPassword" id="confirmPassword" />
-        <button type="submit">Register</button>
-      </form>
+    <div className={styles.registerContainer}>
+      <h2 className={styles.title}>Register</h2>
+      <div className={styles.sectionRegisterForm}>
+        <form onSubmit={registerOnSubmit} className={styles.form}>
+          <div>
+            Full Name : <input type="text" name="fullname" id="fullname" />
+          </div>
+          <div>
+            Email : <input type="email" name="email" id="email" />
+          </div>
+          <div>
+            Password : <input type="password" name="password" id="password" />
+          </div>
+          <div>
+            Confirm Password :{" "}
+            <input
+              type="password"
+              name="confirmPassword"
+              id="confirmPassword"
+            />
+          </div>
+
+          <button type="submit" className={styles.formButton}>
+            Register
+          </button>
+        </form>
+      </div>
     </div>
   );
 };
